Change variable type to evade compilation fail
git-svn-id: svn+ssh://svn.cern.ch/reps/atlasinst/Institutes/UdineICTP/TtHFitter/trunk@383878 28d52da6-859c-4cdc-abcc-1ec3d2c02001

+++ Root/TtHFit.C
@@ -1018,7 +1018,7 @@ void TtHFit::DrawSignalRegionsPlot(int nCols,int nRows, std::vector < Region* > 
         }
     }
     //
-    float yMax = 0;
+    double yMax = 0;
     //
     for(int i=0;i<Nreg;i++){
         pBottom->cd(i+1);
